Refactor dependencies into a unified compilation context `DPL` Fixes #1
